+++ CHANGELOG.md
@@ -7,10 +7,32 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+### Changed
+
+-   Showing a confirmation button before consuming a passwordless link if there is no stored login attempt info or if it doesn't match the link
+-   Showing a confirmation button before consuming an email verification token if there is no active session to prevent mail scanners validating the email address
+-   Adds an example app with Email Verification with OTP
+
+## [0.22.2] - 2022-06-11
+
+### Changes
+
+-   Clearing errors when switching between sign-in and up
+
+### Build changes
+
+-   Updated typescript to latest
+-   Using tsconfig to set jsx import source instead of pragma comments
+-   Updated react-shadow
+
+## [0.22.1] - 2022-06-11
+
+-   Updates the example app for ThirdPartyEmailPassword + Passwordless login with SuperTokens
+-   Adds new tests for testing resend code button in passwordless recipe
+
 ### CI changes
 
 -   Only running React 16 tests on CircleCI (when an explicit envvar is set)
--   Adds an example app with Email Verification with OTP
 
 ## [0.22.0] - 2022-06-03